pyproject.toml
[project]
name = "context-shield"
version = "1.0.0"
description = "Context-aware content moderation OpenEnv environment"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "openenv-core>=0.2.3",
    "pydantic>=2.0",
    "openai>=1.0.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "httpx>=0.27.0",
    "requests>=2.31.0",
]

[project.optional-dependencies]
dev = ["pytest>=7.0", "hypothesis==6.112.0"]

[project.scripts]
server = "server.app:main"

[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["env", "graders", "models", "tasks", "server"]
